chore: put canSwitchToSmartAccount to util
